Initial commit exploring some cases I found where bigendian encoding is enabled, in uncommon, or illogical circumstances.
"Endianness refers to the sequential order in which bytes are arranged into larger numerical values"

My test cases revolve arround situations where we are less than 1 byte, or non even divisions of bytes.

For example, we should Ignore big endian decoding logic for signals with 8 bit or less.

Furthermore, I believe I identified an error when there is bigendian encoding in the last byte of a message, in which an incorrect calculation triggers a failed sanity check of message length